# Optimised the weather component code