**1. The Addressed Issue
I. state_attributes**
An excessive number of unnecessary function calls slowed down the original state_attributes code in the weather component. Key inefficiencies included:

- Repeated calls to unit conversion and commonly used functions (e.g., _default_temperature_unit, _temperature_unit), causing performance bottlenecks.
- These redundancies significantly increased the number of function calls, slowing down the function and impacting overall system performance.

**II. _temperature_unit**
The _temperature_unit property was inefficient due to frequent evaluations of conditions and retrieval of configuration options.

- The lack of a caching mechanism resulted in redundant processing every time the property was accessed, which was frequent, given its central role in the weather component’s operations.
- This inefficiency scaled poorly in scenarios with a high number of calls, resulting in slower performance of the overall system.
- The profile_weather_entity.py script was created to profile and benchmark the performance of key functions, such as state_attributes, within the weather component. It provides detailed insights into execution times and function calls, enabling targeted optimisations to improve efficiency and scalability in real-world scenarios.

The profile_weather_entity.py script was created to profile and benchmark the performance of critical methods like state_attributes in the weather component, ensuring that inefficiencies are identified and addressed effectively. The weather_entity_profile.prof file is the output of this profiling process, capturing detailed performance metrics such as execution times and function calls, which serve as the foundation for implementing targeted optimisations and measuring their impact.

**2. What I Have Reengineered: Code optimisation**
**To address these issues, I use code optimising to reengineer the state_attributes function by:**

- Implemented caching for frequently used values like the temperature unit and precision, ensuring they are computed only once per function call.
- Introduced a helper function, convert_and_round, to centralise unit conversion and rounding logic. This reduced repetitive code and improved readability.
- Streamlined the logic by minimizing redundant calls to properties like _temperature_unit and _default_temperature_unit, leveraging the cached values instead.

After these changes, state_attributes ran faster and did fewer calculations that didn't change during a single run.

**I optimised the unit_temperature function by:**

- The _temperature_unit property was reengineered to incorporate a caching mechanism, reducing redundant evaluations and ensuring the value was computed only once during its lifecycle.
- A lazy evaluation approach was implemented, where the property calculates its value on the first access and caches the result for subsequent accesses.

This change streamlined access to the temperature unit, particularly in scenarios where the property is accessed multiple times in quick succession, such as during weather data updates.

**3. Code Optimization Approach**
**The code optimisation approach for state_attributes:**

- Minimizing Redundant Operations: Reducing repeated calls by caching frequent values within the function using local variables.
- Efficient Function Structure: Consolidating repetitive operations into a single helper function (convert_and_round) makes the code more efficient and easier to follow.
- Streamlining Execution: Grouping related tasks, such as unit conversions, to allow the function to complete its operations with fewer calls and dependencies.

**For _unit_temperature:**

- Lazy Evaluation: The property now computes its value only on first access and stores it for future use, preventing unnecessary recalculations.
- Local Caching: The function avoids redundant logic evaluations in subsequent calls by storing the result in a local variable or attribute.
- Minimising Function Calls: The optimisation reduced dependency on auxiliary functions, which previously contributed to performance overhead.

This approach improved the function’s performance and readability while retaining the original functionality.

**4. Impact of Changes**
**State_attribute:**

- Reduced Function Calls: The total number of function calls dropped from about 31 million to 22 million (around 30% reduction), lowering system load.
- Faster Execution: Execution time improved from 9 seconds to 6.3 seconds (around 25.9% time reduction), demonstrating enhanced function efficiency.
- Improved Code Maintainability: The function is now more structured and readable, making future maintenance easier. Grouped operations and eliminated redundancies make the codebase more organised.

**Screenshots**
[Before](https://drive.google.com/file/d/1kWMKpui2K9TjsvUYR0MdPE4W2eI4K1Xr/view?usp=drive_link)
[After](https://drive.google.com/file/d/1f-HJqmbVhXf_tKyPE5noUy2J9cTGpOE4/view?usp=drive_link)

**_temperature_unit:**

- Reduced execution time: The execution time for the _temperature_unit property decreased from 0.986 seconds to 0.852 seconds, yielding a 13.6% improvement in performance. Contributed to the overall reduction in the execution time of functions like state_attributes and compounded system performance, particularly in scenarios where weather data is updated or accessed frequently.
- The current performance can be accessed visually using ‘snakeviz weather_entity_profile.prof’ in a virtual environment on the terminal.

[Before](https://drive.google.com/file/d/17axgjYQXxj9bHXv6cKIWW1gQ47NPeXTG/view?usp=drive_link)
[After](https://drive.google.com/file/d/1VyyfRAMIkfXQlEWR6SyQn7htGFfOPb72/view?usp=drive_link)